功能更新和健壮性更新
- ComfyUI-Upscale.py的手动模式支持单文件或从外部文件读取文件列表
- 增加：删除重复行.py
- 根据文档复制文件.py在遇到不是文件或文件不存在的行时会提醒
- PNG转JPG.py样式表更新和bug修复
- 增加图片筛选.py

+++ CLI/根据文档复制文件.py
@@ -12,8 +12,13 @@ if not os.path.exists(temp_path):
 
 # function to copy file
 def copy_file(file_name):
-    shutil.copy(file_name, temp_path)
-    print(f"已复制文件：{file_name} 到 {temp_path}")
+    if os.path.exists(file_name):
+        shutil.copy(file_name, temp_path)
+        print(f"已复制文件：{file_name} 到 {temp_path}")
+    elif not os.path.isfile(file_name):
+        print(f"不是文件：{file_name}")
+    else:
+        print(f"文件不存在：{file_name}")
 
 # read every line in txt
 with open(txt_path, 'r', encoding='utf-8') as f:
